// src/app/dropdown/types.ts
/**
 *
 *
 * @export
 * @interface MultiSelectOption
 */
export interface MultiSelectOption {
  id: any;
  label: string;
}


/**
 *
 *
 * @export
 * @interface MultiSelectTexts
 */
export interface MultiSelectTexts {
  checked?: string;
  checkedPlural?: string;
  searchPlaceholder?: string;
  searchEmptyResult?: string;
  defaultTitle?: string;
  allSelected?: string;
}